move DynDevice to its own module
