+++ wgpu-hal/src/dynamic/mod.rs
@@ -1,10 +1,14 @@
 mod command;
+mod device;
+
+pub use self::command::DynCommandEncoder;
+pub use self::device::DynDevice;
 
 use std::any::Any;
 
 use wgt::WasmNotSendSync;
 
-use crate::{BufferBinding, CommandEncoder, Device};
+use crate::BufferBinding;
 
 // TODO: docs
 pub trait DynResource: Any + WasmNotSendSync + 'static {
@@ -75,58 +79,6 @@ pub trait DynQuerySet: DynResource + std::fmt::Debug {}
 pub trait DynRenderPipeline: DynResource + std::fmt::Debug {}
 pub trait DynTexture: DynResource + std::fmt::Debug {}
 pub trait DynTextureView: DynResource + std::fmt::Debug {}
-
-pub trait DynDevice {
-    unsafe fn destroy_buffer(&self, buffer: Box<dyn DynBuffer>);
-}
-
-impl<D: Device> DynDevice for D {
-    unsafe fn destroy_buffer(&self, mut buffer: Box<dyn DynBuffer>) {
-        // Ideally, we'd cast the box and then unbox it with `Box::into_inner`.
-        // Unfortunately, the latter is only available on nightly Rust.
-        //
-        // Another better alternative would be for `D::destroy_buffer` to take a `Box<D::A::Buffer>`.
-        // However, that would require casting the box first to `Box<dyn Any>` for which we need
-        // super trait casting (https://rust-lang.github.io/rfcs/3324-dyn-upcasting.html)
-        // which as of writing is still being stabilized.
-        let buffer = buffer.expect_downcast_mut();
-        unsafe { self.destroy_buffer(buffer) };
-    }
-}
-
-pub trait DynCommandEncoder {
-    unsafe fn set_index_buffer<'a>(
-        &mut self,
-        binding: BufferBinding<'a, dyn DynBuffer>,
-        format: wgt::IndexFormat,
-    );
-
-    unsafe fn set_vertex_buffer<'a>(
-        &mut self,
-        index: u32,
-        binding: BufferBinding<'a, dyn DynBuffer>,
-    );
-}
-
-impl<C: CommandEncoder> DynCommandEncoder for C {
-    unsafe fn set_index_buffer<'a>(
-        &mut self,
-        binding: BufferBinding<'a, dyn DynBuffer>,
-        format: wgt::IndexFormat,
-    ) {
-        let binding = binding.expect_downcast();
-        unsafe { self.set_index_buffer(binding, format) };
-    }
-
-    unsafe fn set_vertex_buffer<'a>(
-        &mut self,
-        index: u32,
-        binding: BufferBinding<'a, dyn DynBuffer>,
-    ) {
-        let binding = binding.expect_downcast();
-        unsafe { self.set_vertex_buffer(index, binding) };
-    }
-}
 
 impl<'a> BufferBinding<'a, dyn DynBuffer> {
     pub fn expect_downcast<B: DynBuffer>(self) -> BufferBinding<'a, B> {
